+++ changelog.md
@@ -15,7 +15,10 @@
   * Food Hunger Multiplier is now 1 and his old value is now applied to 'Food Heal.Health Multiplier'
   * Healing speed is now inverse (high saturation gives slower regen)
   * Fixed raw food not healing
-* Iron veins are now slightly bigger but generate less when exposed o air
+* Iron veins are now slightly bigger but generate less when exposed to air
+* Added World module
+  * Added Desire paths, walking on the same path over and over has a chance to transform grass into dirt and dirt into coarse dirt
+  * Moved loot purger to it
 
 ## Alpha 3.0.0
 * First Alpha